updated .gitignore
set up .gitattributes for selected files

[bin] set all scripts executable

[multigrid_gravity_helper] fixed messages

[multigrid_gravity] Removed local FFT solver as it is dead for too long

Removed prefer_rbgs_relaxation from various problem.par

[multigrid_gravity] removed prefer_rbgs_relaxation parameter

[multigrid_gravity] enabled FFT solver only on the base level

[multigrid_gravity] removed code obsoleted by c30e22dfcede8e0c09b69f251df4bad1b9a7a525

[multigrid_gravity] Improved sanity checks for FFT base level solver

+++ src/multigrid/multigridvars.F90
@@ -122,7 +122,7 @@ module multigridvars
       type(grid_container), pointer, intent(inout) :: cg                     !< current grid container
       integer(kind=4),               intent(in)    :: ind                    !< index in cg%q(:)
       integer,                       intent(out)   :: is, ie, js, je, ks, ke !< indices in cg
-      integer,                       intent(in)    :: b                      !< how far we look into boundary layer
+      integer(kind=4),               intent(in)    :: b                      !< how far we look into boundary layer
       logical,                       intent(in)    :: need_bnd_upd           !< if .true. then update 1 layer of external boundaries
 
       ! calling curl%external_boundaries(ind, bnd_type = BND_NEGREF) is a bit overkill
